Restructure service layer
